requirements.txt
streamlit
langchain
google-generativeai
langchain-google-genai
edge-tts
SpeechRecognition
python-dotenv
pdfplumber
plotly
langgraph
reportlab